implemented algorithm for image conversion

+++ model/src/main/java/edu/kit/ifv/trafficspvisualizer/model/BufferedImageConvertible.java
@@ -3,5 +3,5 @@ package edu.kit.ifv.trafficspvisualizer.model;
 import java.awt.image.BufferedImage;
 
 public interface BufferedImageConvertible {
-    BufferedImage toBufferedImage(int height, int width);
+    BufferedImage toBufferedImage(float height, float width);
 }
